+++ render/wgpu/src/lib.rs
@@ -29,6 +29,7 @@ pub mod target;
 mod uniform_buffer;
 
 pub mod backend;
+mod blend;
 mod buffer_pool;
 #[cfg(feature = "clap")]
 pub mod clap;
